requirements.txt
blinker==1.9.0
click==8.2.1
filelock==3.13.1
Flask==3.1.1
fsspec==2024.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.3
numpy==2.3.1
opencv-python==4.11.0.86
pillow==11.0.0
setuptools==70.2.0
sympy==1.13.3
torch==2.7.1+cpu
torchaudio==2.7.1+cpu
torchvision==0.22.1+cpu
typing_extensions==4.12.2
Werkzeug==3.1.3
